requirements.txt
python-telegram-bot[job-queue]>=22,<23
python-dotenv>=1.0.1,<2
